Navigation + Delete post button

+++ Enigmath/news/views.py
@@ -26,8 +26,12 @@ def news_create(request):
         # message success
         messages.success(request, "Successfully Created")
         return HttpResponseRedirect(instance.get_absolute_url())
+    see_delete_button = "no"
+    if request.user.is_staff or request.user.is_superuser:
+        see_delete_button = "yes"
     context = {
         "form": form,
+        "see_delete_button":see_delete_button,
     }
     return render(request, "news_create.html", context)
 
@@ -42,6 +46,10 @@ def news_detail(request, slug=None):
         "content_type": instance.get_content_type,
         "object_id": instance.id
     }
+
+    see_delete_button = "no"
+    if request.user.is_staff or request.user.is_superuser:
+        see_delete_button = "yes"
 
     form = CommentForm(request.POST or None, initial=initial_data)
 
@@ -77,13 +85,16 @@ def news_detail(request, slug=None):
         "comments": instance.comments,
         "comment_form":form,
         "post_url":instance.get_absolute_url(),
+        "see_delete_button":see_delete_button,
     }
     return render(request, "news_detail.html", context)
 
 def news_list(request):
     today = timezone.now().date()
     queryset_list = Post.objects.active() #.order_by("-timestamp")
+    see_delete_button = "no"
     if request.user.is_staff or request.user.is_superuser:
+        see_delete_button = "yes"
         queryset_list = Post.objects.all()
     
     query = request.GET.get("q")
@@ -112,6 +123,7 @@ def news_list(request):
         "title": "News",
         "page_request_var": page_request_var,
         "today": today,
+        "see_delete_button" : see_delete_button,
     }
     return render(request, "news_list.html", context)
 
@@ -140,9 +152,20 @@ def news_update(request, slug=None):
 
 
 def news_delete(request, slug=None):
-    if not request.user.is_staff or not request.user.is_superuser:
+    try:
+        instance = Post.objects.get(slug=slug)
+    except:
         raise Http404
-    instance = get_object_or_404(Post, slug=slug)
-    instance.delete()
-    messages.success(request, "Successfully deleted")
-    return redirect("news:list")
+
+    if not request.user.is_staff or not request.user.is_superuser:
+        reponse.status_code = 403
+        return HttpResponse("You do not have permission to do this.")
+
+    if request.method == "POST":
+        instance.delete()
+        messages.success(request, "Successfully deleted")
+        return redirect("news:list")
+    context = {
+        "object": instance
+    }
+    return render(request, "confirm_delete.html", context)
